rockchip: Print a message when returning to the bootrom
At present if the return to bootrom fails (e.g. because you are not using
the Rockchip's bootrom's pointer table in MMC) then the board prints
SPL message and hangs. Print a message first if we can, to help in
understanding what happened when it hangs.

Signed-off-by: Simon Glass <sjg@chromium.org>
Series-to: u-boot
Series-cc: Tested-by: Heiko Stuebner <heiko@sntech.de>
Series-version: 2
Series-changes: 2
- Make inclusion of bootrom.c conditional in the Makefile
- Update #ifdef check around printf() to include !CONFIG_TPL_BUILD

Tested-by: Heiko Stuebner <heiko@sntech.de>
Acked-by: Heiko Stuebner <heiko@sntech.de>